# api.yaml
swagger: '2.0'
info:
  title: API микросервиса общих справочников
  version: 1.0.0
paths:
...
  /dictionaries/division-codes:
    get:
      tags:
        - DivisionCodes
      description: Список кодов подразделений
      operationId: listDivisionCodes
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/divisionCode'
            description: Постраничный вывод результатов поиска
    post:
      tags:
        - DivisionCodes
      description: Создание кода подразделения
      operationId: createDivisionCode
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: divisionCode
          description: Данные коде подразделения
          required: true
          schema:
            $ref: '#/definitions/divisionCode'
      responses:
        201:
          description: CREATED
          schema:
            description: Созданный код подразделения
            $ref: '#/definitions/divisionCode'
        409:
          description: Если указан id и кода подразделения с таким id уже существует
  /dictionaries/division-codes/{id}:
    get:
      tags:
        - DivisionCodes
      description: Возвращает информацию о коде подразделения
      operationId: getDivisionCode
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор кода подразделения
      responses:
        200:
          description: ОК
          schema:
            description: Информация о коде подразделения
            $ref: '#/definitions/divisionCode'
        404:
          description: Not Found
  /dictionaries/division-codes#byCode:
    get:
      tags:
        - DivisionCodes
      description: Возвращает информацию о коде подразделения по коду
      operationId: getDivisionCodeByCode
      produces:
        - application/json
      parameters:
        - in: query
          name: code
          type: string
          required: true
          description: Код подразделения
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/divisionCode'
            description: Информация о коде подразделения
        403:
          description: Forbidden
        404:
          description: Not Found

  #***currencies***
...
  /dictionaries/external-systems:
    get:
      tags:
        - ExternalSystems
      description: Список внешних систем
      operationId: listExternalSystems
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/externalSystem'
            description: Постраничный вывод результатов поиска
    post:
      tags:
        - ExternalSystems
      description: Создание внешней системы
      operationId: createExternalSystem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: externalSystem
          description: Данные о внешней системе
          required: true
          schema:
            $ref: '#/definitions/externalSystem'
      responses:
        201:
          description: CREATED
          schema:
            description: Созданная внешняя система
            $ref: '#/definitions/externalSystem'
        409:
          description: Если указан id и внешняя система с таким id уже существует
  /dictionaries/external-systems#bySystemId:
    get:
      tags:
        - ExternalSystems
      description: Поиск внешней системы по идентификатору
      operationId: getExternalSystemBySystemId
      produces:
        - application/json
      parameters:
        - in: query
          name: systemId
          type: string
          description: Идентификатор системы (например "BISQUIT:6300")
          required: false
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/externalSystem'
            description: Внешняя система с заданным кодом
        404:
          description: Not Found
  /dictionaries/external-systems/{id}:
    get:
      tags:
        - ExternalSystems
      description: Возвращает информацию о внешней системе
      operationId: getExternalSystem
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор внешней системы
      responses:
        200:
          description: ОК
          schema:
            description: Информация о внешней системе
            $ref: '#/definitions/externalSystem'
        404:
          description: Not Found
    put:
      tags:
        - ExternalSystems
      description: Редактирование внешней системы
      operationId: editExternalSystem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор внешней системы
        - in: body
          name: externalSystem
          description: Данные о внешней системе
          required: true
          schema:
            $ref: '#/definitions/externalSystem'
      responses:
        200:
          description: ОК
          schema:
            description: Сохраненная внешняя система
            $ref: '#/definitions/externalSystem'
        404:
          description: Not Found
    delete:
      tags:
        - ExternalSystems
      description: Удаление внешней системы
      operationId: deleteExternalSystem
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор внешней системы
      responses:
        204:
          description: No Content
        404:
          description: Not Found


  #***external-system-errors***
  /dictionaries/external-system-errors:
    get:
      tags:
        - ExternalSystemsErrors
      description: Список ошибок внешних систем
      operationId: listExternalSystemsErrors
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/externalSystemError'
            description: Постраничный вывод результатов поиска
    post:
      tags:
        - ExternalSystemsErrors
      description: Создание ошибки внешней системы
      operationId: createExternalSystemsError
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: externalSystem
          description: Данные об ошибке внешней системы
          required: true
          schema:
            $ref: '#/definitions/externalSystemError'
      responses:
        201:
          description: CREATED
          schema:
            description: Созданная ошибка внешней системы
            $ref: '#/definitions/externalSystemError'
        409:
          description: Если указан id и ошибка внешней системы с таким id уже существует
  /dictionaries/external-system-errors/{id}:
    get:
      tags:
        - ExternalSystemsErrors
      description: Возвращает информацию об ошибке внешней системе
      operationId: getExternalSystemsError
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор ошибки внешней системы
      responses:
        200:
          description: ОК
          schema:
            description: Информация об ошибке внешней системы
            $ref: '#/definitions/externalSystemError'
        404:
          description: Not Found
    put:
      tags:
        - ExternalSystemsErrors
      description: Редактирование ошибки внешней системы
      operationId: editExternalSystemsError
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор ошибки внешней системы
        - in: body
          name: externalSystem
          description: Данные об ошибке внешней системы
          required: true
          schema:
            $ref: '#/definitions/externalSystemError'
      responses:
        200:
          description: ОК
          schema:
            description: Сохраненная ошибка внешней системы
            $ref: '#/definitions/externalSystemError'
        404:
          description: Not Found
    delete:
      tags:
        - ExternalSystemsErrors
      description: Удаление ошибки внешней системы
      operationId: deleteExternalSystemsError
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор ошибки внешней системы
      responses:
        204:
          description: No Content
        404:
          description: Not Found

  #***offices***
...
definitions:
...
  divisionCode:
    type: object
    description: Коды подразделений
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      code:
        type: string
        description: код
      description:
        type: string
        description: описание
  externalSystem:
    type: object
    description: Внешняя система
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      name:
        type: string
        description: Название
      systemType:
        type: string
        description: Тип системы
      systemId:
        type: string
        description: Идентификатор системы вида BISQUIT:6300
      connectTimeout:
        type: integer
        format: int64
        description: Таймаут соединения
      failurePeriod:
        type: integer
        format: int64
        description: Период отказа
      failurePeriodCount:
        type: integer
        format: int64
        description: Количество периодов отказа
  externalSystemError:
    type: object
    description: Ошибка внешней системы
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      serviceName:
        type: string
        description: Название внешней системы
      errorCode:
        type: string
        description: Код ошибки
      message:
        type: string
        description: Сообщение об ошибке
...