# common-dict-app/src/main/java/ru/softlab/efr/common/dict/controllers/DivisionCodesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ru.softlab.efr.common.dict.controllers;

import ru.softlab.efr.common.dict.exchange.model.DivisionCode;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.validation.BindingResult;
import org.springframework.data.web.PageableDefault;
import org.springframework.data.domain.*;

import java.util.*;

import org.springframework.validation.annotation.Validated;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * the DivisionCodes API
 */
public interface DivisionCodesApi {

    /**
     * Создание кода подразделения 
     * @param divisionCode Данные коде подразделения  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/division-codes",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<DivisionCode> createDivisionCode(@Valid @RequestBody DivisionCode divisionCode) throws Exception;



    /**
     * Возвращает информацию о коде подразделения 
     * @param id Идентификатор кода подразделения  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/division-codes/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<DivisionCode> getDivisionCode(@PathVariable("id") Long id) throws Exception;



    /**
     * Возвращает информацию о коде подразделения по коду 
     * @param code Код подразделения  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/division-codes#byCode",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<DivisionCode> getDivisionCodeByCode(@NotNull  @Valid @RequestParam(value = "code") String code) throws Exception;



    /**
     * Список кодов подразделений   
     * @param pageable параметры пагинации 
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/division-codes",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Page<DivisionCode>> listDivisionCodes(@PageableDefault(value = 50) Pageable pageable) throws Exception;


}