# common-dict-app/src/main/java/ru/softlab/efr/common/dict/controllers/ExternalSystemsErrorsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ru.softlab.efr.common.dict.controllers;

import ru.softlab.efr.common.dict.exchange.model.ExternalSystemError;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.validation.BindingResult;
import org.springframework.data.web.PageableDefault;
import org.springframework.data.domain.*;

import java.util.*;

import org.springframework.validation.annotation.Validated;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * the ExternalSystemsErrors API
 */
public interface ExternalSystemsErrorsApi {

    /**
     * Создание ошибки внешней системы 
     * @param externalSystem Данные об ошибке внешней системы  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/external-system-errors",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<ExternalSystemError> createExternalSystemsError(@Valid @RequestBody ExternalSystemError externalSystem) throws Exception;



    /**
     * Удаление ошибки внешней системы 
     * @param id Идентификатор ошибки внешней системы  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/external-system-errors/{id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteExternalSystemsError(@PathVariable("id") Long id) throws Exception;



    /**
     * Редактирование ошибки внешней системы 
     * @param id Идентификатор ошибки внешней системы
     * @param externalSystem Данные об ошибке внешней системы  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/external-system-errors/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<ExternalSystemError> editExternalSystemsError(@PathVariable("id") Long id,@Valid @RequestBody ExternalSystemError externalSystem) throws Exception;



    /**
     * Возвращает информацию об ошибке внешней системе 
     * @param id Идентификатор ошибки внешней системы  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/external-system-errors/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<ExternalSystemError> getExternalSystemsError(@PathVariable("id") Long id) throws Exception;



    /**
     * Список ошибок внешних систем   
     * @param pageable параметры пагинации 
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/external-system-errors",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Page<ExternalSystemError>> listExternalSystemsErrors(@PageableDefault(value = 50) Pageable pageable) throws Exception;


}